some tests changes mostly
