+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/FeedToken/ChangeFeedIteratorCoreTests.cs
@@ -16,6 +16,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Microsoft.Azure.Cosmos.CosmosElements;
     using Newtonsoft.Json;
+    using Newtonsoft.Json.Linq;
 
     [SDK.EmulatorTests.TestClass]
     [TestCategory("ChangeFeed")]
@@ -49,6 +50,22 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
         {
             ContainerResponse response = await this.database.CreateContainerAsync(
                 new ContainerProperties(id: Guid.NewGuid().ToString(), partitionKeyPath: ChangeFeedIteratorCoreTests.PartitionKey),
+                cancellationToken: this.cancellationToken);
+
+            return (ContainerInternal)response;
+        }
+
+        private async Task<ContainerInternal> InitializeFFCFContainerAsync(TimeSpan timeToLive)
+        {
+            ContainerProperties containerProperties = new(id: Guid.NewGuid().ToString(), partitionKeyPath: @"/id")
+            {
+                DefaultTimeToLive = Convert.ToInt32(timeToLive.TotalSeconds),
+            };
+
+            containerProperties.ChangeFeedPolicy.FullFidelityRetention = TimeSpan.FromMinutes(5);
+
+            ContainerResponse response = await this.database.CreateContainerAsync(
+                containerProperties,
                 cancellationToken: this.cancellationToken);
 
             return (ContainerInternal)response;
@@ -129,7 +146,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
             DateTime now = DateTime.UtcNow;
             await Task.Delay(1000);
             await this.CreateRandomItems(itemsCore, batchSize, randomPartitionKey: true);
-            
+
             FeedIterator feedIterator = itemsCore.GetChangeFeedStreamIterator(
                 ChangeFeedStartFrom.Time(now),
                 ChangeFeedMode.Incremental);
@@ -342,7 +359,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
 
             ContainerInternal itemsCore = await this.InitializeContainerAsync();
             await this.CreateRandomItems(itemsCore, batchSize, randomPartitionKey: true);
-            
+
             FeedIterator<ToDoActivity> feedIterator = itemsCore.GetChangeFeedIterator<ToDoActivity>(ChangeFeedStartFrom.Beginning(), ChangeFeedMode.Incremental);
             string continuation = null;
             while (feedIterator.HasMoreResults)
@@ -433,7 +450,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
         {
             ContainerInternal itemsCore = await this.InitializeContainerAsync();
             await this.CreateRandomItems(itemsCore, 2, randomPartitionKey: true);
-            
+
             ChangeFeedIteratorCore feedIterator = itemsCore.GetChangeFeedStreamIterator(
                 ChangeFeedStartFrom.Beginning(),
                 ChangeFeedMode.Incremental,
@@ -478,7 +495,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
             int expected = 25;
             int iterations = 0;
             await this.CreateRandomItems(itemsCore, expected, randomPartitionKey: true);
-            
+
             string continuation = null;
             int count = 0;
             while (true)
@@ -537,7 +554,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
             CosmosObject previousToken = null;
             ContainerInternal itemsCore = await this.InitializeLargeContainerAsync();
             await this.CreateRandomItems(itemsCore, expected, randomPartitionKey: true);
-            
+
             ChangeFeedIteratorCore feedIterator = itemsCore.GetChangeFeedStreamIterator(
                 ChangeFeedStartFrom.Beginning(),
                 ChangeFeedMode.Incremental,
@@ -580,7 +597,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
         {
             ContainerInternal itemsCore = await this.InitializeLargeContainerAsync();
             int pkRangesCount = (await itemsCore.ClientContext.DocumentClient.ReadPartitionKeyRangeFeedAsync(itemsCore.LinkUri)).Count;
-            
+
             IEnumerable<FeedRange> tokens = await itemsCore.GetFeedRangesAsync();
             Assert.AreEqual(pkRangesCount, tokens.Count());
         }
@@ -590,7 +607,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
         {
             ContainerInternal itemsCore = await this.InitializeLargeContainerAsync();
             int pkRangesCount = (await itemsCore.ClientContext.DocumentClient.ReadPartitionKeyRangeFeedAsync(itemsCore.LinkUri)).Count;
-            
+
             IEnumerable<FeedRange> tokens = await itemsCore.GetFeedRangesAsync();
             Assert.IsTrue(pkRangesCount > 1, "Should have created a multi partition container.");
             Assert.AreEqual(pkRangesCount, tokens.Count());
@@ -781,6 +798,280 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
             Assert.IsTrue(cosmosException.Message.Contains("FullFidelity Change Feed must have valid If-None-Match header."));
         }
 
+        /// <summary>
+        /// This test will execute <see cref="Container.GetChangeFeedIterator{T}(ChangeFeedStartFrom, ChangeFeedMode, ChangeFeedRequestOptions)"/> in <see cref="ChangeFeedMode.OperationsLog"/> (FullFidelity) with a typed item.
+        /// Using FeedRange.FromPartitionKey.
+        /// </summary>
+        [TestMethod]
+        public async Task ChangeFeedIteratorCore_FeedRange_FromPartitionKey_VerifyingWireFormatTests()
+        {
+            TimeSpan ttl = TimeSpan.FromSeconds(1);
+            ContainerInternal container = await this.InitializeFFCFContainerAsync(ttl);
+            string id = Guid.NewGuid().ToString();
+            string otherId = Guid.NewGuid().ToString();
+            using (FeedIterator<ChangeFeedItemChanges<Item>> feedIterator = container.GetChangeFeedIterator<ChangeFeedItemChanges<Item>>(
+                changeFeedStartFrom: ChangeFeedStartFrom.Now(FeedRange.FromPartitionKey(new PartitionKey(id))),
+                changeFeedMode: ChangeFeedMode.FullFidelity))
+            {
+                string continuation = null;
+                while (feedIterator.HasMoreResults)
+                {
+                    FeedResponse<ChangeFeedItemChanges<Item>> feedResponse = await feedIterator.ReadNextAsync();
+
+                    if (feedResponse.StatusCode == HttpStatusCode.NotModified)
+                    {
+                        continuation = feedResponse.ContinuationToken;
+                        Assert.IsNotNull(continuation);
+
+                        PartitionKey partitionKey = new(id);
+                        PartitionKey otherPartitionKey = new(otherId);
+
+                        _ = await container.UpsertItemAsync<Item>(item: new(Id: otherId, Line1: "87 38floor, Witthayu Rd, Lumphini, Pathum Wan District", City: "Bangkok", State: "Thailand", ZipCode: "10330"), partitionKey: otherPartitionKey).ConfigureAwait(false);
+                        _ = await container.UpsertItemAsync<Item>(item: new(Id: id, Line1: "One Microsoft Way", City: "Redmond", State: "WA", ZipCode: "98052"), partitionKey: partitionKey).ConfigureAwait(false);
+                        _ = await container.UpsertItemAsync<Item>(item: new(Id: id, Line1: "205 16th St NW", City: "Atlanta", State: "GA", ZipCode: "30363"), partitionKey: partitionKey).ConfigureAwait(false);
+                    }
+                    else
+                    {
+#if DEBUG
+                        Console.WriteLine(JsonConvert.SerializeObject(feedResponse.Resource));
+#endif
+                        List<ChangeFeedItemChanges<Item>> itemChanges = feedResponse.Resource.ToList();
+
+                        ChangeFeedIteratorCoreTests.AssertGatewayMode(feedResponse);
+
+                        Assert.AreEqual(expected: 2, actual: itemChanges.Count);
+
+                        ChangeFeedItemChanges<Item> createOperation = itemChanges[0];
+
+                        Assert.AreEqual(expected: id, actual: createOperation.Current.Id);
+                        Assert.AreEqual(expected: "One Microsoft Way", actual: createOperation.Current.Line1);
+                        Assert.AreEqual(expected: "Redmond", actual: createOperation.Current.City);
+                        Assert.AreEqual(expected: "WA", actual: createOperation.Current.State);
+                        Assert.AreEqual(expected: "98052", actual: createOperation.Current.ZipCode);
+                        Assert.IsNotNull(createOperation.Metadata);
+                        Assert.AreEqual(expected: ChangeFeedOperationType.Create, actual: createOperation.Metadata.OperationType);
+                        Assert.AreNotEqual(notExpected: default, actual: createOperation.Metadata.ConflictResolutionTimestamp);
+                        Assert.AreNotEqual(notExpected: default, actual: createOperation.Metadata.CurrentLogSequenceNumber);
+                        Assert.AreEqual(expected: default, actual: createOperation.Metadata.PreviousLogSequenceNumber);
+                        Assert.IsFalse(createOperation.Metadata.TimeToLiveExpired);
+
+                        ChangeFeedItemChanges<Item> replaceOperation = itemChanges[1];
+
+                        Assert.AreEqual(expected: id, actual: replaceOperation.Current.Id);
+                        Assert.AreEqual(expected: "205 16th St NW", actual: replaceOperation.Current.Line1);
+                        Assert.AreEqual(expected: "Atlanta", actual: replaceOperation.Current.City);
+                        Assert.AreEqual(expected: "GA", actual: replaceOperation.Current.State);
+                        Assert.AreEqual(expected: "30363", actual: replaceOperation.Current.ZipCode);
+                        Assert.IsNotNull(createOperation.Metadata);
+                        Assert.AreEqual(expected: ChangeFeedOperationType.Replace, actual: replaceOperation.Metadata.OperationType);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.ConflictResolutionTimestamp);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.CurrentLogSequenceNumber);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.PreviousLogSequenceNumber);
+                        Assert.IsFalse(replaceOperation.Metadata.TimeToLiveExpired);
+
+                        break;
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// This test will execute <see cref="Container.GetChangeFeedIterator{T}(ChangeFeedStartFrom, ChangeFeedMode, ChangeFeedRequestOptions)"/> in <see cref="ChangeFeedMode.OperationsLog"/> (FullFidelity) with a typed item.
+        /// Using ChangeFeedStartFrom.Now(ranges[0]).
+        /// </summary>
+        [TestMethod]
+        public async Task ChangeFeedIteratorCore_FeedRange_VerifyingWireFormatTests()
+        {
+            TimeSpan ttl = TimeSpan.FromSeconds(1);
+            ContainerInternal container = await this.InitializeFFCFContainerAsync(ttl);
+            IReadOnlyList<FeedRange> ranges = await container.GetFeedRangesAsync();
+            string id = Guid.NewGuid().ToString();
+            string otherId = Guid.NewGuid().ToString();
+
+            using (FeedIterator<ChangeFeedItemChanges<Item>> feedIterator = container.GetChangeFeedIterator<ChangeFeedItemChanges<Item>>(
+                changeFeedStartFrom: ChangeFeedStartFrom.Now(ranges[0]),
+                changeFeedMode: ChangeFeedMode.FullFidelity))
+            {
+                string continuation = null;
+                while (feedIterator.HasMoreResults)
+                {
+                    FeedResponse<ChangeFeedItemChanges<Item>> feedResponse = await feedIterator.ReadNextAsync();
+
+                    if (feedResponse.StatusCode == HttpStatusCode.NotModified)
+                    {
+                        continuation = feedResponse.ContinuationToken;
+                        Assert.IsNotNull(continuation);
+
+                        PartitionKey partitionKey = new(id);
+                        PartitionKey otherPartitionKey = new(otherId);
+
+                        _ = await container.UpsertItemAsync<Item>(item: new(Id: otherId, Line1: "87 38floor, Witthayu Rd, Lumphini, Pathum Wan District", City: "Bangkok", State: "Thailand", ZipCode: "10330"), partitionKey: otherPartitionKey).ConfigureAwait(false);
+                        _ = await container.UpsertItemAsync<Item>(item: new(Id: id, Line1: "One Microsoft Way", City: "Redmond", State: "WA", ZipCode: "98052"), partitionKey: partitionKey).ConfigureAwait(false);
+                        _ = await container.UpsertItemAsync<Item>(item: new(Id: id, Line1: "205 16th St NW", City: "Atlanta", State: "GA", ZipCode: "30363"), partitionKey: partitionKey).ConfigureAwait(false);
+                        _ = await container.DeleteItemAsync<Item>(id: id, partitionKey: partitionKey);
+                    }
+                    else
+                    {
+#if DEBUG
+                        Console.WriteLine(JsonConvert.SerializeObject(feedResponse.Resource));
+#endif
+                        List<ChangeFeedItemChanges<Item>> resources = feedResponse.Resource.ToList();
+
+                        ChangeFeedIteratorCoreTests.AssertGatewayMode(feedResponse);
+
+                        Assert.AreEqual(expected: 4, actual: resources.Count);
+
+                        ChangeFeedItemChanges<Item> firstCreateOperation = resources[0];
+
+                        Assert.AreEqual(expected: otherId, actual: firstCreateOperation.Current.Id);
+                        Assert.AreEqual(expected: "87 38floor, Witthayu Rd, Lumphini, Pathum Wan District", actual: firstCreateOperation.Current.Line1);
+                        Assert.AreEqual(expected: "Bangkok", actual: firstCreateOperation.Current.City);
+                        Assert.AreEqual(expected: "Thailand", actual: firstCreateOperation.Current.State);
+                        Assert.AreEqual(expected: "10330", actual: firstCreateOperation.Current.ZipCode);
+                        Assert.AreEqual(expected: ChangeFeedOperationType.Create, actual: firstCreateOperation.Metadata.OperationType);
+                        Assert.AreNotEqual(notExpected: default, actual: firstCreateOperation.Metadata.ConflictResolutionTimestamp);
+                        Assert.AreNotEqual(notExpected: default, actual: firstCreateOperation.Metadata.CurrentLogSequenceNumber);
+                        Assert.AreEqual(expected: default, actual: firstCreateOperation.Metadata.PreviousLogSequenceNumber);
+                        Assert.IsFalse(firstCreateOperation.Metadata.TimeToLiveExpired);
+
+                        ChangeFeedItemChanges<Item> createOperation = resources[1];
+
+                        Assert.AreEqual(expected: id, actual: createOperation.Current.Id);
+                        Assert.AreEqual(expected: "One Microsoft Way", actual: createOperation.Current.Line1);
+                        Assert.AreEqual(expected: "Redmond", actual: createOperation.Current.City);
+                        Assert.AreEqual(expected: "WA", actual: createOperation.Current.State);
+                        Assert.AreEqual(expected: "98052", actual: createOperation.Current.ZipCode);
+                        Assert.AreEqual(expected: ChangeFeedOperationType.Create, actual: createOperation.Metadata.OperationType);
+                        Assert.AreNotEqual(notExpected: default, actual: createOperation.Metadata.ConflictResolutionTimestamp);
+                        Assert.AreNotEqual(notExpected: default, actual: createOperation.Metadata.CurrentLogSequenceNumber);
+                        Assert.AreEqual(expected: default, actual: createOperation.Metadata.PreviousLogSequenceNumber);
+                        Assert.IsFalse(createOperation.Metadata.TimeToLiveExpired);
+
+                        ChangeFeedItemChanges<Item> replaceOperation = resources[2];
+
+                        Assert.AreEqual(expected: id, actual: replaceOperation.Current.Id);
+                        Assert.AreEqual(expected: "205 16th St NW", actual: replaceOperation.Current.Line1);
+                        Assert.AreEqual(expected: "Atlanta", actual: replaceOperation.Current.City);
+                        Assert.AreEqual(expected: "GA", actual: replaceOperation.Current.State);
+                        Assert.AreEqual(expected: "30363", actual: replaceOperation.Current.ZipCode);
+                        Assert.AreEqual(expected: ChangeFeedOperationType.Replace, actual: replaceOperation.Metadata.OperationType);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.ConflictResolutionTimestamp);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.CurrentLogSequenceNumber);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.PreviousLogSequenceNumber);
+                        Assert.IsFalse(replaceOperation.Metadata.TimeToLiveExpired);
+
+                        ChangeFeedItemChanges<Item> deleteOperation = resources[3];
+
+                        Assert.IsNull(deleteOperation.Current.Id);
+                        Assert.IsNull(deleteOperation.Current.Line1);
+                        Assert.IsNull(deleteOperation.Current.City);
+                        Assert.IsNull(deleteOperation.Current.State);
+                        Assert.IsNull(deleteOperation.Current.ZipCode);
+                        Assert.AreEqual(expected: ChangeFeedOperationType.Delete, actual: deleteOperation.Metadata.OperationType);
+                        Assert.AreNotEqual(notExpected: default, actual: deleteOperation.Metadata.ConflictResolutionTimestamp);
+                        Assert.AreNotEqual(notExpected: default, actual: deleteOperation.Metadata.CurrentLogSequenceNumber);
+                        Assert.AreNotEqual(notExpected: default, actual: deleteOperation.Metadata.PreviousLogSequenceNumber);
+                        Assert.IsNotNull(deleteOperation.Previous);
+                        Assert.AreEqual(expected: id, actual: deleteOperation.Previous.Id);
+                        Assert.AreEqual(expected: "205 16th St NW", actual: deleteOperation.Previous.Line1);
+                        Assert.AreEqual(expected: "Atlanta", actual: deleteOperation.Previous.City);
+                        Assert.AreEqual(expected: "GA", actual: deleteOperation.Previous.State);
+                        Assert.AreEqual(expected: "30363", actual: deleteOperation.Previous.ZipCode);
+
+                        break;
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// This test will execute <see cref="Container.GetChangeFeedIterator{T}(ChangeFeedStartFrom, ChangeFeedMode, ChangeFeedRequestOptions)"/> in <see cref="ChangeFeedMode.OperationsLog"/> (FullFidelity) with a typed item.
+        /// Using FeedRange.FromPartitionKey.
+        /// </summary>
+        [TestMethod]
+        public async Task ChangeFeedIteratorCore_FeedRange_FromPartitionKey_Dynamic_VerifyingWireFormatTests()
+        {
+            TimeSpan ttl = TimeSpan.FromSeconds(1);
+            ContainerInternal container = await this.InitializeFFCFContainerAsync(ttl);
+            string id = Guid.NewGuid().ToString();
+            string otherId = Guid.NewGuid().ToString();
+            using (FeedIterator<dynamic> feedIterator = container.GetChangeFeedIterator<dynamic>(
+                changeFeedStartFrom: ChangeFeedStartFrom.Now(FeedRange.FromPartitionKey(new PartitionKey(id))),
+                changeFeedMode: ChangeFeedMode.FullFidelity))
+            {
+                string continuation = null;
+                while (feedIterator.HasMoreResults)
+                {
+                    FeedResponse<dynamic> feedResponse = await feedIterator.ReadNextAsync();
+
+                    if (feedResponse.StatusCode == HttpStatusCode.NotModified)
+                    {
+                        continuation = feedResponse.ContinuationToken;
+                        Assert.IsNotNull(continuation);
+
+                        PartitionKey partitionKey = new(id);
+                        PartitionKey otherPartitionKey = new(otherId);
+
+                        _ = await container.UpsertItemAsync<dynamic>(item: new { id = otherId, line1 = "87 38floor, Witthayu Rd, Lumphini, Pathum Wan District", city = "Bangkok", state = "Thailand", zipCode = "10330" }, partitionKey: otherPartitionKey).ConfigureAwait(false);
+                        _ = await container.UpsertItemAsync<dynamic>(item: new { id, line1 = "One Microsoft Way", city = "Redmond", state = "WA", zipCode = "98052" }, partitionKey: partitionKey).ConfigureAwait(false);
+                        _ = await container.UpsertItemAsync<dynamic>(item: new { id, line1 = "205 16th St NW", city = "Atlanta", state = "GA", zipCode = "30363" }, partitionKey: partitionKey).ConfigureAwait(false);
+                        _ = await container.DeleteItemAsync<Item>(id: id, partitionKey: partitionKey);
+                    }
+                    else
+                    {
+#if DEBUG
+                        Console.WriteLine(JsonConvert.SerializeObject(feedResponse.Resource));
+#endif
+                        List<ChangeFeedItemChanges<Item>> itemChanges = JsonConvert.DeserializeObject<List<ChangeFeedItemChanges<Item>>>(
+                            JsonConvert.SerializeObject(feedResponse.Resource));
+
+                        ChangeFeedIteratorCoreTests.AssertGatewayMode(feedResponse);
+
+                        Assert.AreEqual(expected: 2, actual: itemChanges.Count);
+
+                        ChangeFeedItemChanges<Item> createOperation = itemChanges[0];
+
+                        Assert.AreEqual(expected: id, actual: createOperation.Current.Id);
+                        Assert.AreEqual(expected: "One Microsoft Way", actual: createOperation.Current.Line1);
+                        Assert.AreEqual(expected: "Redmond", actual: createOperation.Current.City);
+                        Assert.AreEqual(expected: "WA", actual: createOperation.Current.State);
+                        Assert.AreEqual(expected: "98052", actual: createOperation.Current.ZipCode);
+                        Assert.IsNotNull(createOperation.Metadata);
+                        Assert.AreEqual(expected: ChangeFeedOperationType.Create, actual: createOperation.Metadata.OperationType);
+                        Assert.AreNotEqual(notExpected: default, actual: createOperation.Metadata.ConflictResolutionTimestamp);
+                        Assert.AreNotEqual(notExpected: default, actual: createOperation.Metadata.CurrentLogSequenceNumber);
+                        Assert.AreEqual(expected: default, actual: createOperation.Metadata.PreviousLogSequenceNumber);
+                        Assert.IsFalse(createOperation.Metadata.TimeToLiveExpired);
+
+                        ChangeFeedItemChanges<Item> replaceOperation = itemChanges[1];
+
+                        Assert.AreEqual(expected: id, actual: replaceOperation.Current.Id);
+                        Assert.AreEqual(expected: "205 16th St NW", actual: replaceOperation.Current.Line1);
+                        Assert.AreEqual(expected: "Atlanta", actual: replaceOperation.Current.City);
+                        Assert.AreEqual(expected: "GA", actual: replaceOperation.Current.State);
+                        Assert.AreEqual(expected: "30363", actual: replaceOperation.Current.ZipCode);
+                        Assert.IsNotNull(createOperation.Metadata);
+                        Assert.AreEqual(expected: ChangeFeedOperationType.Replace, actual: replaceOperation.Metadata.OperationType);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.ConflictResolutionTimestamp);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.CurrentLogSequenceNumber);
+                        Assert.AreNotEqual(notExpected: default, actual: replaceOperation.Metadata.PreviousLogSequenceNumber);
+                        Assert.IsFalse(replaceOperation.Metadata.TimeToLiveExpired);
+
+                        break;
+                    }
+                }
+            }
+        }
+
+        private static void AssertGatewayMode<T>(FeedResponse<T> feedResponse)
+        {
+            string diagnostics = feedResponse.Diagnostics.ToString();
+            JToken jsonToken = JToken.Parse(diagnostics);
+
+            Assert.IsNotNull(jsonToken["Summary"]["GatewayCalls"], "'GatewayCalls' is not found in diagnostics. UseGateMode is set to false.");
+        }
+
+
         private async Task<IList<ToDoActivity>> CreateRandomItems(ContainerInternal container, int pkCount, int perPKItemCount = 1, bool randomPartitionKey = true)
         {
             Assert.IsFalse(!randomPartitionKey && perPKItemCount > 1);
@@ -821,7 +1112,7 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
 
         private class CancellationTokenRequestHandler : RequestHandler
         {
-            public CancellationToken LastUsedToken { get; private set;  }
+            public CancellationToken LastUsedToken { get; private set; }
 
             public override Task<ResponseMessage> SendAsync(RequestMessage request, CancellationToken cancellationToken)
             {
@@ -830,4 +1121,11 @@ namespace Microsoft.Azure.Cosmos.EmulatorTests.FeedRanges
             }
         }
     }
+    
+    public record Item(
+        [property: JsonProperty("id")] string Id,
+        [property: JsonProperty("line1")] string Line1,
+        [property: JsonProperty("city")] string City,
+        [property: JsonProperty("zipCode")] string ZipCode,
+        [property: JsonProperty("state")] string State);
 }
